moved express app to nx workspace
