--- src/settings.json
@@ -1,4 +1,0 @@
-{
-  "energyCosts": 29,
-  "currency": "EUR"
-}
