+++ tplink-energy-monitor/apps/api-server/src/app/models/devices/tp-link-plug-info.dto.ts
