+++ tplink-energy-monitor/apps/api-server/src/app/models/mapper/map-sysinfo-to-tp-link-plug-info.mapper.ts
@@ -1,6 +1,5 @@
 import {
     DeviceEnergyOverview,
-    DeviceMonthlyEnergyOverview,
     DeviceRealtimeData,
     TpLinkPlugInfoDto,
 } from '../devices/tp-link-plug-info.dto';
