+++ tplink-energy-monitor/apps/api-server/src/app/models/mapper/map-to-tp-link-plug.mapper.ts
