+++ tplink-energy-monitor/apps/api-server/src/app/services/devices.service.ts
