+++ tplink-energy-monitor/apps/api-server/src/app/services/file.service.ts
