+++ tplink-energy-monitor/apps/api-server/src/app/types/devices/device-type.enum.ts
