+++ tplink-energy-monitor/apps/api-server/src/app/types/http-exception.type.ts
