+++ tplink-energy-monitor/package.json
@@ -6,9 +6,13 @@
   "author": {
     "name": "James Barnett"
   },
-  "contributors": {
+  "contributors": [{
     "name": "SebastianSiebert"
-  },
+  },{
+    "name": "xeroxx"
+  },{
+    "name": "nbyx"
+  }],
   "scripts": {
     "nx": "nx",
     "start": "nx serve",
@@ -44,13 +48,19 @@
     "@nrwl/workspace": "11.5.2",
     "@types/express": "4.17.0",
     "@types/jest": "26.0.8",
+    "@types/morgan": "^1.9.2",
     "@types/node": "14.14.33",
+    "@types/socket.io": "^2.1.13",
     "@typescript-eslint/eslint-plugin": "4.3.0",
     "@typescript-eslint/parser": "4.3.0",
     "dotenv": "8.2.0",
     "eslint": "7.10.0",
+    "eslint-config-airbnb-base": "^14.2.1",
     "eslint-config-prettier": "8.1.0",
+    "eslint-plugin-import": "^2.22.1",
+    "eslint-plugin-prettier": "^3.3.1",
     "jest": "26.2.2",
+    "nodemon": "^2.0.7",
     "prettier": "2.2.1",
     "ts-jest": "26.4.0",
     "ts-node": "~9.1.1",
@@ -58,7 +68,15 @@
   },
   "_id": "tplink-energy-monitor@0.0.0",
   "dependencies": {
+    "@overnightjs/core": "^1.7.5",
+    "@overnightjs/logger": "^1.2.0",
+    "body-parser": "^1.19.0",
     "express": "4.17.1",
-    "tslib": "^2.0.0"
+    "http-status-codes": "^2.1.4",
+    "morgan": "^1.10.0",
+    "socket.io": "^3.1.2",
+    "tplink-smarthome-api": "^3.3.0",
+    "tslib": "^2.0.0",
+    "typedi": "^0.8.0"
   }
 }
